osf_demo_01/requirements.txt
greenlet==2.0.2
mysql-connector==2.2.9
mysql-connector-python==8.0.33
protobuf==3.20.3
PyMySQL==1.0.3
python-dotenv==1.0.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3